requirements.txt
asgiref==3.7.2
bleach==5.0.1
channels==4.0.0
channels-redis==4.1.0
crispy-bootstrap5==0.7
Django==4.2.5
django-autoslug==1.9.9
django-bleach==3.1.0
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-htmx==1.16.0
django-js-asset==2.1.0
django-tinymce==3.6.1
django-uploader==0.2.1
install==1.3.5
msgpack==1.0.7
Pillow==10.0.0
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tinycss2==1.1.1
tzdata==2023.3
webencodings==0.5.1
